requirements.txt
numpy
autograd
cyanure-mkl==0.21.post2
scikit-learn
pandas
scipy
logging-python
